Refactor: change project structure

+++ src/test/java/xyz/npgw/test/common/base/BaseTest.java
@@ -15,6 +15,7 @@ import org.testng.annotations.BeforeMethod;
 import org.testng.annotations.Optional;
 import org.testng.annotations.Parameters;
 import xyz.npgw.test.common.BrowserFactory;
+import xyz.npgw.test.common.Constants;
 import xyz.npgw.test.common.PlaywrightOptions;
 import xyz.npgw.test.common.ProjectProperties;
 import xyz.npgw.test.common.ProjectUtils;
@@ -26,7 +27,9 @@ import java.nio.file.Path;
 import java.nio.file.Paths;
 
 public abstract class BaseTest {
+
     protected static final Logger LOGGER = LogManager.getLogger(BaseTest.class.getName());
+
     private static final String ARTEFACT_DIR = "target/artefact";
     private Playwright playwright;
     private Browser browser;
@@ -43,6 +46,9 @@ public abstract class BaseTest {
             browser = BrowserFactory.getBrowser(playwright, this.browserType);
         } catch (IllegalArgumentException e) {
             LOGGER.error("Unsupported browser: {}", browserType);
+            if (playwright != null) {
+                playwright.close();
+            }
             System.exit(1);
         } catch (RuntimeException e) {
             LOGGER.error("Playwright.create() failed: {}", e.getMessage());
@@ -60,40 +66,71 @@ public abstract class BaseTest {
 
         page = context.newPage();
         ProjectUtils.navigateToBaseURL(page);
+
+        if (!method.getDeclaringClass().getSimpleName().contains("LoginPageTest")) {
+            Allure.step("Login to the site");
+            ProjectUtils.login(page);
+            if (page.url().contains(Constants.DASHBOARD_PAGE_URL)) {
+                LOGGER.debug("Dashboard Page is opened");
+            } else {
+                LOGGER.error("Failed to open Dashboard Page. Check your credentials.");
+            }
+        } else {
+            Allure.step("Navigate to the base url");
+            ProjectUtils.navigateToBaseURL(page);
+        }
+
+
     }
 
     @AfterMethod(alwaysRun = true)
     protected void afterMethod(Method method, ITestResult testResult) {
-        page.close();
-
         String testName = ProjectUtils.getTestClassCompleteMethodName(method, testResult);
         Path traceFilePath = Paths.get(ARTEFACT_DIR, browserType, testName + ".zip");
-        if (ProjectProperties.isTracingMode()) {
-            context.tracing().stop(PlaywrightOptions.tracingStopOptions(traceFilePath));
+        Path videoFilePath = Paths.get(ARTEFACT_DIR, browserType, testName + ".webm");
+
+        if (page != null) {
+            page.close();
+
+            if (ProjectProperties.isVideoMode()) {
+                page.video().saveAs(videoFilePath);
+                page.video().delete();
+            }
         }
 
-        Path videoFilePath = Paths.get(ARTEFACT_DIR, browserType, testName + ".webm");
-        if (ProjectProperties.isVideoMode()) {
-            page.video().saveAs(videoFilePath);
-            page.video().delete();
+        if (context != null) {
+            if (ProjectProperties.isTracingMode()) {
+                context.tracing().stop(PlaywrightOptions.tracingStopOptions(traceFilePath));
+            }
+
+            context.close();
         }
 
         if (!testResult.isSuccess()) { // && ProjectProperties.isServerRun()) {
             try {
-                Allure.getLifecycle().addAttachment("video", "video/webm", "webm", Files.readAllBytes(videoFilePath));
-                Allure.getLifecycle().addAttachment("tracing", "archive/zip", "zip", Files.readAllBytes(traceFilePath));
+                if (ProjectProperties.isTracingMode()) {
+                    Allure.getLifecycle().addAttachment(
+                            "video", "video/webm", "webm", Files.readAllBytes(videoFilePath));
+                }
+                if (ProjectProperties.isTracingMode()) {
+                    Allure.getLifecycle().addAttachment(
+                            "tracing", "archive/zip", "zip", Files.readAllBytes(traceFilePath));
+                }
             } catch (IOException e) {
                 LOGGER.error("Add artefacts to allure failed: {}", e.getMessage());
             }
         }
-
-        context.close();
     }
 
     @AfterClass(alwaysRun = true)
     protected void afterClass() {
-        browser.close();
-        playwright.close();
+        if (browser != null) {
+            browser.close();
+        }
+
+        if (playwright != null) {
+            playwright.close();
+        }
     }
 
     protected Page getPage() {
